+++ tests/tests/es.weak-map.js
@@ -65,6 +65,10 @@ QUnit.test('WeakMap', assert => {
     object = {};
     assert.same(new Subclass().set(object, 2).get(object), 2, 'correct subclassing with native classes #3');
   }
+
+  const buffer = new ArrayBuffer(8);
+  const map = new WeakMap([[buffer, 8]]);
+  assert.ok(map.has(buffer), 'works with ArrayBuffer keys');
 });
 
 QUnit.test('WeakMap#delete', assert => {